Switching to a proprietary license.
After this commit and the next commit, the bc-wallet-common-lib-vault repository will evolve and be developed independently of the crypto-bundle project. The license changes to proprietary.

The MIT license is changed by me (Kotelnikov AV) as an author and maintainer.

The license will be replaced with a proprietary one, with the condition that the authorship is preserved and indicated in the README.md file in the section of authors and contributors.

Aleksei Kotelnikov - author and maintainer of crypto-bundle project(https://github.com/crypto-bundle)

The commit is signed with the key -
gudron2s@gmail.com
E456BB23A18A9347E952DBC6655133DD561BF3EC